+++ src/expandableCalendar/calendarProvider.js
@@ -33,7 +33,7 @@ class CalendarProvider extends Component {
       date: this.props.date || XDate().toString('yyyy-MM-dd'),
       updateSource: UPDATE_SOURCES.CALENDAR_INIT,
       buttonY: new Animated.Value(-65),
-      buttonIcon: props.showTodayButton && this.getButtonIcon(props.date),
+      buttonIcon: this.getButtonIcon(props.date),
       disabled: false,
       opacity: new Animated.Value(1)
     };
@@ -55,7 +55,7 @@ class CalendarProvider extends Component {
   };
 
   setDate = (date, updateSource) => {
-    this.setState({date, updateSource, buttonIcon: this.props.showTodayButton && this.getButtonIcon(date)}, () => {
+    this.setState({date, updateSource, buttonIcon: this.getButtonIcon(date)}, () => {
       this.animateTodayButton(date);
     });
     _.invoke(this.props, 'onDateChanged', date, updateSource);
@@ -69,6 +69,9 @@ class CalendarProvider extends Component {
   }
 
   getButtonIcon(date) {
+    if (!this.props.showTodayButton) {
+      return;
+    }
     const isPastDate = this.isPastDate(date);
     return isPastDate ? iconDown : iconUp;
   }